Multimedia03 and Pai update

+++ oopProj/src/main/java/com/oopProj/Repository/TaskRepository.java
@@ -12,8 +12,8 @@ import java.util.List;
 public interface TaskRepository extends JpaRepository<Task, Integer> {
 
     @Query("SELECT z FROM Task z WHERE z.project.projectId = :projectId")
-    Page<Task> findTasksForProject(@Param("projectId") Long projectId, Pageable pageable);
+    Page<Task> findTasksForProject(@Param("projectId") Integer projectId, Pageable pageable);
 
     @Query("SELECT z FROM Task z WHERE z.project.projectId = :projectId")
-    List<Task> findTasksForProject(@Param("projectId") Long projectId);
+    List<Task> findTasksForProject(@Param("projectId") Integer projectId);
 }
